NF: move ChangeDeckMulti where it's used
